[cdatapack] use int for string length
Summary: Unless `PY_SSIZE_T_CLEAN` is defined, the length is int.

Test Plan: compiles.

Reviewers: #fastmanifest, durham

Reviewed By: durham

Subscribers: mitrandir

Differential Revision: https://phabricator.intern.facebook.com/D3666998

Signature: t1:3666998:1470341329:3ddd1f7e36389aff3e18482db4d0b28ab8d7c12f

+++ remotefilelog/cdatapack/py-cdatapack.c
@@ -45,7 +45,7 @@ static int cdatapack_init(py_cdatapack *self, PyObject *args) {
   self->handle = NULL;
 
   char *node;
-  ssize_t nodelen;
+  int nodelen;
 
   if (!PyArg_ParseTuple(args, "s#", &node, &nodelen)) {
     return -1;
